Added more util functions
